Ensure service CIDRs are updated in any case
As NCP does not handle service CIDRs, if these are the only
item being updated on the network CR, NCP skips procesing
the resource status update.

This patch ensures the network CR is updated if necessary
even if the operator is not applying any change to NCP.

Finally, this patch renames the getNetworkCRD function to
buildNetworkStatus as it reflects better what the function
actually does.